db not working pls fix

+++ LocMessClient/app/src/main/java/pt/ulisboa/tecnico/cmu/tg14/locmessclient/DataObjects/Location.java
@@ -63,7 +63,7 @@ public class Location {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
